Avoid integer overflow in CalcPlaygoDigests. Close #53

+++ LibOrbisPkg/PKG/PkgBuilder.cs
@@ -194,7 +194,7 @@ namespace LibOrbisPkg.PKG
           (chunk, _, local) =>
           {
             var (sha, buffer) = local;
-            view.ReadArray(chunk * CHUNK_SIZE, buffer, 0, CHUNK_SIZE);
+            view.ReadArray((long)chunk * CHUNK_SIZE, buffer, 0, CHUNK_SIZE);
             Buffer.BlockCopy(sha.ComputeHash(buffer), 0, FileData, chunk * 4, 4);
             return local;
           },
